main

+++ src/algebra/plus.hpp
@@ -4,6 +4,6 @@ template < class T > class PLUS {
     static constexpr T op(const T &l, const T &r) { return l + r; }
     static constexpr T id = T(0);
     static constexpr T inv(const T &x) { return -x; }
-    static constexpr T pow(const T &x, const int n) { return x * n; }
+    static constexpr T pow(const T &x, const ll n) { return x * n; }
     static constexpr bool comm = true;
 };
